--- a/.travis.yml
+++ b/.travis.yml
@@ -29,7 +29,9 @@ jobs:
     - name: macOS | x86_64 | Python 3.x
       os: osx
       osx_image: xcode10.2
-      env: PYTHON=python3
+      env:
+        - PYTHON=python3
+        - PYTEST_ADDOPTS='-k "unit_test or test_0_basic" --suppress-no-test-exit-code'
 
     - name: Windows | x86_64 | Python 3.6
       os: windows
@@ -53,6 +55,6 @@ jobs:
     # must repeat the s390x job above exactly to match
     - *linux_s390x_36
 
-install: $PYTHON -m pip install -r requirements-dev.txt
+install: $PYTHON -m pip install -r requirements-dev.txt pytest-custom-exit-code
 
 script: $PYTHON ./bin/run_tests.py