--- a/appveyor.yml
+++ b/appveyor.yml
@@ -1,7 +1,4 @@
 environment:
-  # workaround an appveyor bug on macos
-  # see https://github.com/appveyor/ci/issues/3376#issuecomment-611314299
-  APPVEYOR_CONSOLE_DISABLE_PTY: true
   matrix:
     - APPVEYOR_BUILD_WORKER_IMAGE: Ubuntu
       APPVEYOR_JOB_NAME: "python37-x64-ubuntu"