--- a/cibuildwheel/resources/pinned_docker_images.cfg
+++ b/cibuildwheel/resources/pinned_docker_images.cfg
@@ -9,7 +9,7 @@ manylinux2010 = quay.io/pypa/manylinux2010_i686:2020-09-20-6131e60
 manylinux2014 = quay.io/pypa/manylinux2014_i686:2020-09-20-fb9af88
 
 [pypy_x86_64]
-manylinux2010 = yannickjadoul/manylinux2010-pypy_x86_64:7.3.2rc2
+manylinux2010 = yannickjadoul/manylinux2010-pypy_x86_64:7.3.2
 
 [aarch64]
 manylinux2014 = quay.io/pypa/manylinux2014_aarch64:2020-09-20-fb9af88