--- a/cibuildwheel/resources/pinned_docker_images.cfg
+++ b/cibuildwheel/resources/pinned_docker_images.cfg
@@ -1,22 +1,22 @@
 [x86_64]
-manylinux1 = quay.io/pypa/manylinux1_x86_64:2020-09-20-9444d5c
-manylinux2010 = quay.io/pypa/manylinux2010_x86_64:2020-09-20-6131e60
-manylinux2014 = quay.io/pypa/manylinux2014_x86_64:2020-09-20-fb9af88
+manylinux1 = quay.io/pypa/manylinux1_x86_64:2020-09-26-9f151ab
+manylinux2010 = quay.io/pypa/manylinux2010_x86_64:2020-09-26-58cbbcd
+manylinux2014 = quay.io/pypa/manylinux2014_x86_64:2020-09-26-7f7a14f
 
 [i686]
-manylinux1 = quay.io/pypa/manylinux1_i686:2020-09-20-9444d5c
-manylinux2010 = quay.io/pypa/manylinux2010_i686:2020-09-20-6131e60
-manylinux2014 = quay.io/pypa/manylinux2014_i686:2020-09-20-fb9af88
+manylinux1 = quay.io/pypa/manylinux1_i686:2020-09-26-9f151ab
+manylinux2010 = quay.io/pypa/manylinux2010_i686:2020-09-26-58cbbcd
+manylinux2014 = quay.io/pypa/manylinux2014_i686:2020-09-26-7f7a14f
 
 [pypy_x86_64]
-manylinux2010 = yannickjadoul/manylinux2010-pypy_x86_64:7.3.2
+manylinux2010 = pypywheels/manylinux2010-pypy_x86_64:2020-09-26-8f210b5
 
 [aarch64]
-manylinux2014 = quay.io/pypa/manylinux2014_aarch64:2020-09-20-fb9af88
+manylinux2014 = quay.io/pypa/manylinux2014_aarch64:2020-09-26-7f7a14f
 
 [ppc64le]
-manylinux2014 = quay.io/pypa/manylinux2014_ppc64le:2020-09-20-fb9af88
+manylinux2014 = quay.io/pypa/manylinux2014_ppc64le:2020-09-26-7f7a14f
 
 [s390x]
-manylinux2014 = quay.io/pypa/manylinux2014_s390x:2020-09-20-fb9af88
+manylinux2014 = quay.io/pypa/manylinux2014_s390x:2020-09-26-7f7a14f