--- a/examples/circleci-minimal.yml
+++ b/examples/circleci-minimal.yml
@@ -19,7 +19,8 @@ jobs:
   osx-wheels:
     working_directory: ~/osx-wheels
     macos:
-      xcode: "10.0.0"
+      # PyPy 7.3.2 needs macOS >= 10.14
+      xcode: 10.2.1
     steps:
       - checkout
       - run: