--- a/examples/travis-ci-minimal.yml
+++ b/examples/travis-ci-minimal.yml
@@ -15,6 +15,8 @@ jobs:
       arch: s390x
     # and a mac build
     - os: osx
+      # PyPy 7.3.2 needs macOS >= 10.14
+      osx_image: xcode10.2
       language: shell
     # and a windows build
     - os: windows