Currently this is still `rc1`, but let's try out if things work. Only to be merged after the official final release.