@YannickJadoul 2020-09-10
Not sure why Travis CI's macOS job is acting up/so slow, but apart from that, all green :-)