@joerick 2020-09-11
Great! Looks like the Appveyor macOS build is working again, too. Could we remove the expected failure from that CI file?