@YannickJadoul 2020-09-18
Our mac builds on Travis CI are really slow, it seems