@YannickJadoul 2020-09-26
One step closer. But let's wait for pypy/manylinux#10? Smaller images are great for everyone :-)