@YannickJadoul 2020-09-26
If PyPy wasn't suddenly failing on macOS, this would be ready ... :-(