@YannickJadoul 2020-09-26
Logged an issue with PyPy: https://foss.heptapod.net/pypy/pypy/-/issues/3311