@YannickJadoul 2020-09-26
> Logged an issue with PyPy: https://foss.heptapod.net/pypy/pypy/-/issues/3311

Got some feedback on IRC:

> ```
> <mattip> YannickJadoul: what version of MacOS are you testing?
> <YannickJadoul> mattip: CircleCI is running 10.13.3, it seems
> <YannickJadoul> Ah, and yes, Travis as well: macos10.13-xcode9.4.1
> <YannickJadoul> Wikipedia announces 10.13's EOL as "November 30, 2020"
> <mattip> ok, so that is the problem. The buildbot updated xcode, and they must have broken something.
> <mattip> YannickJadoul: for CircleCI see https://circleci.com/docs/2.0/testing-ios/#supported-xcode-versions
> <YannickJadoul> Yes, to confirm, AppVeyor runs 10.14, Azure & GitHub Actions run 10.15. Those jobs succeed
> <YannickJadoul> So should we just update our tests and consider 10.14 the minimum for PyPy to run?
> <mattip> that is one approach, but then all your wheels will be 10.14+, which might not be good for cibuildwheel's users
> <mattip> I added that flag to tonights build, and if it runs to completion will merge to the py3.6, py3.7 branches tomorrow
> <mattip> so maybe you could try a nightly build if those builds succeed
> <YannickJadoul> Will they be? I'm not entirely sure; we do set this MACOSX_DEPLOYMENT_TARGET env-var to 10.9
> <YannickJadoul> OK, great! I don't think we're in a hurry to merge this branch, so I can test with the nightlies tomorrow
> ```

See https://foss.heptapod.net/pypy/pypy/-/commit/30c1deda5676eca5368bd1bcc3a843ad4b0f0612

So we'll can try tomorrow, and otherwise, we can just update our macOS versions on the CI :-)