@YannickJadoul 2020-09-27
> So we'll can try tomorrow, and otherwise, we can just update our macOS versions on the CI :-)

OK, so that didn't work. Updating to macOS 10.14 to get PyPy working.

@joerick, @Czaki, is this safe? Both Travis CI as well as CircleCI just have newer images (but 10.13 is still the default on Travis). If a solution is found and released in PyPy 7.3.3, we can always still revert this change. But 10.13 is also almost end-of-life (Wikipedia says end of November 2020), so it might not be worth bothering?