@joerick 2020-10-02
> @joerick, @Czaki, is this safe? Both Travis CI as well as CircleCI just have newer images (but 10.13 is still the default on Travis). If a solution is found and released in PyPy 7.3.3, we can always still revert this change. But 10.13 is also almost end-of-life (Wikipedia says end of November 2020), so it might not be worth bothering?

I don't see a problem updating our macOS image. It's a little boring that we have to add another line to each example config file, but seems pretty reasonable to me!