@YannickJadoul 2020-10-02
> It's a little boring that we have to add another line to each example config file, but seems pretty reasonable to me!

Oh, right, I forgot to update the examples! Will do!

It's fixed already in PyPy, btw (see [PyPy issue 3311](https://foss.heptapod.net/pypy/pypy/-/issues/3311)), but we'd have to wait for 7.3.3 to revert. I think I prefer to get this in and to finally have all CIs build successfully again?

Also, Travis CI's macOS job keeps on timing out. I guess we should limit the amount of tests, similar to AppVeyor?