@YannickJadoul 2020-10-03
Added the macOS 10.14 images to the examples for Travis CI and CircleCI, and made Travis CI run less tests (this is very minimal, though, should we run more?)

If tests now pass, this should to be ready, I believe. And then we'll be waiting for 7.3.3 :-)