@joerick 2020-10-04
I'm wondering if there's something more going on regarding test timings here - on the master branch the macOS github runner takes 32min to finish - here it's taking 77min, more than double!

I've tried to find the tests that are causing that difference - here's some figures (the Github runners only):

test | master | yannickjadoul:pypy-7.3.2 | time increase
-- | -- | -- | --
test/test_0_basic.py::test | 478.05 | 1217.28 | 739.23
test/test_before_build.py::test_cwd | 99.14 | 390.49 | 291.35
test/test_before_build.py::test | 96.90 | 375.37 | 278.47
test/test_before_test.py::test | 266.06 | 493.51 | 227.45
test/test_before_all.py::test_cwd | 95.26 | 306.23 | 210.97
test/test_before_all.py::test | 96.38 | 284.62 | 188.24
test/test_cpp_standards.py::test_cpp11 | 89.68 | 135.08 | 45.4
test/test_cpp_standards.py::test_cpp14 | 84.25 | 118.76 | 34.51
test/test_dependency_versions.py::test_pinned_versions[2.7] | 23.13 | 33.28 | 10.15
test/test_before_build.py::test_failing_command | 6.86 | 16.79 | 9.93
test/test_ssl.py::test | 108.67 | 116.23 | 7.56
test/test_dependency_versions.py::test_pinned_versions[3.6] | 29.89 | 36.80 | 6.91
test/test_0_basic.py::test_build_identifiers | 0.15 | 2.14 | 1.99
test/test_pure_wheel.py::test | 5.71 | 7.24 | 1.53
test/test_testing.py::test_failing_test | 9.30 | 10.58 | 1.28
test/test_before_all.py::test_failing_command | 0.57 | 1.36 | 0.79
test/test_dependency_versions.py::test_dependency_constraints_file[2.7] | 31.88 | 32.10 | 0.22
test/test_environment.py::test_overridden_path | 1.68 | 0.27 | -1.41
test/test_dependency_versions.py::test_pinned_versions[3.8] | 15.27 | 9.27 | -6
test/test_testing.py::test_extras_require | 167.80 | 161.48 | -6.32
test/test_dependency_versions.py::test_pinned_versions[3.5] | 17.83 | 11.21 | -6.62
test/test_testing.py::test | 179.22 | 171.56 | -7.66
test/test_subdir_package.py::test | 21.85 | 11.61 | -10.24
test/test_pep518.py::test_pep518 | 181.32 | 170.25 | -11.07
test/test_cpp_standards.py::test_cpp17 | 69.64 | 51.42 | -18.22
test/test_dependency_versions.py::test_dependency_constraints_file[3.x] | 127.67 | 107.49 | -20.18
test/test_build_skip.py::test | 56.73 | 34.93 | -21.8
test/test_environment.py::test | 171.42 | 129.80 | -41.62

Most notable is that first `test_0_basic`, where all the versions of Python are installed for the first time. Could indicate slow downloads? Then again, the azure macOS runner is finishing fast enough - 36 minutes. So I'm pretty confused about the whole thing. Still, if there's some issue that's affecting GithubCI and Travis, but not Azure and CircleCI, it would be good to find it out!