@Czaki 2020-10-04
Maybe setup cache?
And it looks like unacceptable slow of build.