@YannickJadoul 2020-10-04
> I'm wondering if there's something more going on regarding test timings here

Huh, I hadn't considered that, but you seem to be right :-/

> Most notable is that first `test_0_basic`, where all the versions of Python are installed for the first time. Could indicate slow downloads?

Relatively speaking, `test_before_build` is even worse, though. Times have sometimes almost quadrupled, while `test_0_basic` "only" +- doubled?

Could we try reverting the changes from this PR one by one, until this doesn't happen anymore?