@YannickJadoul 2020-10-04
> Could we try reverting the changes from this PR one by one, until this doesn't happen anymore?

Happy to do so, but I'd like to make sure @joerick isn't testing other things :-)