@joerick 2020-10-04
Looks like times are back to normal! I'm wondering if it was some kind of network event that was causing slow builds, e.g. slow Python download CDN or a slow PyPI.

Maybe we should try removing the Travis single-test thing on Travis now as well? I wonder if it might work again now.