@YannickJadoul 2020-10-04
Sure, maybe just `git revert 0154a8f99207e8d727f5d420ccb3486d4ae43e47`?